Change "needle" prop to "requestLib"
Should never use a specific library name as a property like that. Needle could be replaced with any lib that supports the basic get/post/patch/delete verbs (e.g. request) and it should not affect the DC class API. Instead default it to needle internally but name it generically on the class.

Also fixed the indentation on the "crawls" level of unit tests